ai_processing/tests.py
"""
Essential tests for the ai_processing app.

Tests cover core functionality:
- API endpoints
- Service classes (sentiment, summary, speech)
- Model operations  
- Basic utility functions
"""
import tempfile
from unittest.mock import Mock, patch
from django.test import TestCase
from django.contrib.auth.models import User
from rest_framework.test import APITestCase
from rest_framework import status
from .models import AIProcessingResult


class AIProcessingModelTest(TestCase):
    """Test the AIProcessingResult model"""
    
    def setUp(self):
        self.user = User.objects.create_user(username='testuser', password='testpass')
    
    def test_create_processing_result(self):
        """Test creating an AI processing result"""
        result = AIProcessingResult.objects.create(
            user=self.user,
            processing_type='sentiment_analysis',
            input_text='Test message',
            result_data={'sentiment': 'positive'},
            status='completed',
            processing_time=1.5
        )
        
        self.assertEqual(result.processing_type, 'sentiment_analysis')
        self.assertEqual(result.status, 'completed')
        self.assertTrue(result.is_successful)


class SentimentAnalysisTest(TestCase):
    """Test sentiment analysis service"""
    
    def test_analyse_positive_sentiment(self):
        """Test analyzing positive sentiment with mock service"""
        # Since we don't want to import actual service classes that may have dependencies,
        # we'll test the concept with a mock
        with patch('ai_processing.services.SentimentAnalysisService') as mock_service:
            mock_instance = Mock()
            mock_instance.analyse_sentiment.return_value = {'sentiment': 'positive', 'confidence': 0.8}
            mock_service.return_value = mock_instance
            
            from ai_processing.services import SentimentAnalysisService
            service = SentimentAnalysisService()
            result = service.analyse_sentiment("I love working on this project!")
            
            self.assertIsNotNone(result)
            self.assertEqual(result['sentiment'], 'positive')
    
    def test_analyse_empty_text(self):
        """Test analyzing empty text returns None"""
        with patch('ai_processing.services.SentimentAnalysisService') as mock_service:
            mock_instance = Mock()
            mock_instance.analyse_sentiment.return_value = None
            mock_service.return_value = mock_instance
            
            from ai_processing.services import SentimentAnalysisService
            service = SentimentAnalysisService()
            result = service.analyse_sentiment("")
            
            self.assertIsNone(result)


class SummaryGenerationTest(TestCase):
    """Test summary generation service"""
    
    def test_generate_summary_with_context(self):
        """Test generating summary with context"""
        with patch('ai_processing.services.SummaryGenerationService') as mock_service:
            mock_instance = Mock()
            mock_instance.generate_summary.return_value = "Test AI summary"
            mock_service.return_value = mock_instance
            
            from ai_processing.services import SummaryGenerationService
            service = SummaryGenerationService()
            
            context = {
                'standup_content': 'Did some coding work',
                'user_info': {'username': 'test'}
            }
            
            result = service.generate_summary(context)
            self.assertEqual(result, "Test AI summary")
    
    def test_generate_summary_no_context(self):
        """Test generating summary without context returns None"""
        with patch('ai_processing.services.SummaryGenerationService') as mock_service:
            mock_instance = Mock()
            mock_instance.generate_summary.return_value = None
            mock_service.return_value = mock_instance
            
            from ai_processing.services import SummaryGenerationService
            service = SummaryGenerationService()
            result = service.generate_summary({})
            
            self.assertIsNone(result)


class SpeechToTextTest(TestCase):
    """Test speech-to-text service"""
    
    def test_transcribe_audio_success(self):
        """Test successful audio transcription"""
        with patch('ai_processing.services.SpeechToTextService') as mock_service:
            mock_instance = Mock()
            mock_instance.transcribe_audio.return_value = 'Hello world'
            mock_service.return_value = mock_instance
            
            from ai_processing.services import SpeechToTextService
            service = SpeechToTextService()
            
            # Create a temporary file to simulate audio
            with tempfile.NamedTemporaryFile(suffix='.wav') as temp_file:
                result = service.transcribe_audio(temp_file.name)
                self.assertEqual(result, 'Hello world')
    
    def test_transcribe_nonexistent_file(self):
        """Test transcribing nonexistent file returns None"""
        with patch('ai_processing.services.SpeechToTextService') as mock_service:
            mock_instance = Mock()
            mock_instance.transcribe_audio.return_value = None
            mock_service.return_value = mock_instance
            
            from ai_processing.services import SpeechToTextService
            service = SpeechToTextService()
            result = service.transcribe_audio('/nonexistent/file.wav')
            
            self.assertIsNone(result)


class AIProcessingAPITest(APITestCase):
    """Test AI processing API endpoints"""
    
    def setUp(self):
        self.user = User.objects.create_user(username='testuser', password='testpass')
        self.client.force_authenticate(user=self.user)
    
    def test_process_standup_text(self):
        """Test processing standup text via API"""
        data = {
            'text_update': 'Yesterday I worked on the dashboard',
            'context': '{"project": "test"}'  # JSON string instead of dict
        }
        response = self.client.post('/api/v1/ai/process/', data, format='json')
        
        # Should return some response (success or handled error)
        self.assertIn(response.status_code, [200, 400, 404, 500])
    
    def test_process_empty_request(self):
        """Test processing empty request"""
        response = self.client.post('/api/v1/ai/process/', {})
        # Should handle empty request gracefully
        self.assertIn(response.status_code, [400, 404])
    
    def test_transcribe_endpoint_exists(self):
        """Test transcription endpoint accessibility"""
        response = self.client.post('/api/v1/ai/speech/transcribe/')
        # Should be accessible (even if it fails due to missing data)
        # 404 means URL doesn't exist, anything else means it's accessible
        self.assertNotEqual(response.status_code, 404)
    
    def test_standup_parse_endpoint_exists(self):
        """Test standup parsing endpoint accessibility"""
        response = self.client.post('/api/v1/ai/standup/parse/', {})
        # Should be accessible (even if it fails due to missing data)
        self.assertNotEqual(response.status_code, 404)
    
    def test_standup_parse_with_transcription(self):
        """Test standup parsing with valid transcription"""
        import json
        transcription = "Yesterday I worked on the API endpoints. Today I will focus on the frontend. I'm blocked by the database issue."
        response = self.client.post('/api/v1/ai/standup/parse/', 
            data=json.dumps({'transcription': transcription}),
            content_type='application/json')
        
        self.assertEqual(response.status_code, 200)
        data = response.json()
        self.assertIn('yesterday', data)
        self.assertIn('today', data)
        self.assertIn('blockers', data)
    
    def test_standup_parse_empty_transcription(self):
        """Test standup parsing with empty transcription"""
        import json
        response = self.client.post('/api/v1/ai/standup/parse/', 
            data=json.dumps({'transcription': ''}),
            content_type='application/json')
        
        self.assertEqual(response.status_code, 400)
    
    def test_standup_parse_no_transcription(self):
        """Test standup parsing without transcription field"""
        import json
        response = self.client.post('/api/v1/ai/standup/parse/', 
            data=json.dumps({}),
            content_type='application/json')
        self.assertEqual(response.status_code, 400)


class StandupParsingServiceTest(TestCase):
    """Test the StandupParsingService"""
    
    def setUp(self):
        from .services import StandupParsingService
        self.service = StandupParsingService()
    
    def test_parse_empty_transcription(self):
        """Test parsing empty transcription returns empty fields"""
        result = self.service.parse_standup_transcription("")
        expected = {"yesterday": "", "today": "", "blockers": ""}
        self.assertEqual(result, expected)
    
    def test_parse_structured_transcription(self):
        """Test parsing a well-structured transcription"""
        transcription = "Yesterday I completed the user authentication module. Today I will work on the dashboard interface. I'm blocked by waiting for the API documentation."
        result = self.service.parse_standup_transcription(transcription)
        
        self.assertIn("authentication", result["yesterday"].lower())
        self.assertIn("dashboard", result["today"].lower())
        self.assertIn("api documentation", result["blockers"].lower())
    
    def test_parse_simple_transcription(self):
        """Test parsing a simple transcription with fallback logic"""
        transcription = "I need to finish the project by Friday."
        result = self.service.parse_standup_transcription(transcription)
        
        # Should use fallback and put it in today
        self.assertEqual(result["yesterday"], "")
        self.assertIn("finish", result["today"].lower())
        self.assertEqual(result["blockers"], "")